+++ ving/docs/change-log.md
@@ -29,6 +29,7 @@ With this:
 Also don't forget to update the imports in your schema files.
 
 * Set filterQualifier to true for the baseSchemaId.
+* Added better error handling for int2str parseId().
 
 ## December 2024